tidy up app.js
