+++ server/src/app.js
@@ -11,7 +11,6 @@ import publicRoute from './routes/publicRoute.js';
 
 import logger from './lib/logger.js';
 import { removeOldImages, webcamWrapper } from './services/camService.js';
-import { checkForErrors, removeOldData, updateKeys } from './services/stationService.js';
 import { soundingWrapper } from './services/soundingService.js';
 
 const app = express();
@@ -40,95 +39,7 @@ cron.schedule('*/10 * * * *', async () => {
   logger.info(`--- Update webcams end - ${Date.now() - ts}ms elapsed.`, { service: 'cam' });
 });
 
-// stations
-// cron.schedule('*/10 * * * *', async () => {
-//   logger.info('--- Update stations start ---', { service: 'station', type: 'other' });
-//   const ts = Date.now();
-//   await stationWrapper();
-//   logger.info(`--- Update stations end - ${Date.now() - ts}ms elapsed.`, {
-//     service: 'station',
-//     type: 'other'
-//   });
-// });
-// cron.schedule('*/10 * * * *', async () => {
-//   logger.info('--- Update harvest stations start ---', { service: 'station', type: 'harvest' });
-//   const ts = Date.now();
-//   await stationWrapper('harvest');
-//   logger.info(`--- Update harvest stations end - ${Date.now() - ts}ms elapsed.`, {
-//     service: 'station',
-//     type: 'harvest'
-//   });
-// });
-// cron.schedule('*/10 * * * *', async () => {
-//   logger.info('--- Update metservice stations start ---', {
-//     service: 'station',
-//     type: 'metservice'
-//   });
-//   const ts = Date.now();
-//   await stationWrapper('metservice');
-//   logger.info(`--- Update metservice stations end - ${Date.now() - ts}ms elapsed.`, {
-//     service: 'station',
-//     type: 'metservice'
-//   });
-// });
-// cron.schedule('*/10 * * * *', async () => {
-//   logger.info('--- Update holfuy stations start ---', { service: 'station', type: 'holfuy' });
-//   const ts = Date.now();
-//   await holfuyWrapper();
-//   logger.info(`--- Update holfuy stations end - ${Date.now() - ts}ms elapsed.`, {
-//     service: 'station',
-//     type: 'holfuy'
-//   });
-// });
-// cron.schedule('*/2 * * * *', async () => {
-//   logger.info('--- Update high resolution stations start ---', { service: 'station', type: 'hr' });
-//   const ts = Date.now();
-//   await highResolutionStationWrapper();
-//   logger.info(`--- Update high resolution stations end - ${Date.now() - ts}ms elapsed.`, {
-//     service: 'station',
-//     type: 'hr'
-//   });
-// });
-// cron.schedule('5,15,25,35,45,55 * * * *', async () => {
-//   logger.info('--- Check missed readings start ---', { service: 'miss' });
-//   const ts = Date.now();
-//   await checkForMissedReadings();
-//   logger.info(`--- Check missed readings end - ${Date.now() - ts}ms elapsed.`, {
-//     service: 'miss'
-//   });
-// });
-
-// json
-// cron.schedule('2,12,22,32,42,52 * * * *', async () => {
-//   logger.info('--- Process json output start ---', { service: 'json' });
-//   const ts = Date.now();
-//   await jsonOutputWrapper();
-//   logger.info(`--- Process json output end - ${Date.now() - ts}ms elapsed.`, { service: 'json' });
-// });
-
-// errors
-cron.schedule('5 */6 * * *', async () => {
-  logger.info('--- Check errors start ---', { service: 'errors' });
-  const ts = Date.now();
-  await checkForErrors();
-  logger.info(`--- Check errors end - ${Date.now() - ts}ms elapsed.`, { service: 'errors' });
-});
-
-// keys
-cron.schedule('5 0 * * *', async () => {
-  logger.info('--- Update keys start ---', { service: 'keys' });
-  const ts = Date.now();
-  await updateKeys();
-  logger.info(`--- Update keys end - ${Date.now() - ts}ms elapsed.`, { service: 'keys' });
-});
-
 // cleanup
-cron.schedule('5 0 * * *', async () => {
-  logger.info('--- Remove old data start ---', { service: 'cleanup' });
-  const ts = Date.now();
-  await removeOldData();
-  logger.info(`--- Remove old data end - ${Date.now() - ts}ms elapsed.`, { service: 'cleanup' });
-});
 cron.schedule('5 0 * * *', async () => {
   logger.info('--- Remove old images start ---', { service: 'cleanup' });
   const ts = Date.now();
